% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.Study}
\alias{autoplot.Study}
\title{Plot an roptuna study}
\usage{
\method{autoplot}{Study}(object, type = "history", ...)
}
\arguments{
\item{object}{A \code{Study} R6 object.}

\item{type}{One of \code{"history"}, \code{"parallel_coordinate"}, \code{"param_importance"}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot an roptuna study
}
\examples{
study <- create_study(direction = "minimize", sampler = tpe_sampler(seed = 1L))
study$optimize(function(trial) {
  x <- trial$suggest_float("x", -5, 5)
  y <- trial$suggest_int("y", 1L, 3L)
  x^2 + y
}, n_trials = 10)
ggplot2::autoplot(study, type = "history")
}
