% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{categorical_distribution}
\alias{categorical_distribution}
\title{Create a categorical distribution}
\usage{
categorical_distribution(choices)
}
\arguments{
\item{choices}{Character or numeric vector of choices.}
}
\value{
An S3 object of class \code{roptuna_categorical_distribution}.
}
\description{
Create a categorical distribution
}
\examples{
categorical_distribution(c("adam", "sgd", "rmsprop"))
}
