% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-cmaes.R
\name{CmaEsSampler}
\alias{CmaEsSampler}
\title{CMA-ES sampler R6 class}
\description{
See \code{\link[=cmaes_sampler]{cmaes_sampler()}} for the recommended constructor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CmaEsSampler-new}{\code{CmaEsSampler$new()}}
\item \href{#method-CmaEsSampler-infer_relative_search_space}{\code{CmaEsSampler$infer_relative_search_space()}}
\item \href{#method-CmaEsSampler-sample_relative}{\code{CmaEsSampler$sample_relative()}}
\item \href{#method-CmaEsSampler-sample_independent}{\code{CmaEsSampler$sample_independent()}}
\item \href{#method-CmaEsSampler-clone}{\code{CmaEsSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmaEsSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-CmaEsSampler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmaEsSampler$new(n_startup_trials, sigma0, seed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmaEsSampler-infer_relative_search_space"></a>}}
\if{latex}{\out{\hypertarget{method-CmaEsSampler-infer_relative_search_space}{}}}
\subsection{Method \code{infer_relative_search_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmaEsSampler$infer_relative_search_space(study, trial)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmaEsSampler-sample_relative"></a>}}
\if{latex}{\out{\hypertarget{method-CmaEsSampler-sample_relative}{}}}
\subsection{Method \code{sample_relative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmaEsSampler$sample_relative(study, trial, search_space, all_params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmaEsSampler-sample_independent"></a>}}
\if{latex}{\out{\hypertarget{method-CmaEsSampler-sample_independent}{}}}
\subsection{Method \code{sample_independent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmaEsSampler$sample_independent(study, trial, param_name, dist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmaEsSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CmaEsSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmaEsSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
