% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{float_distribution}
\alias{float_distribution}
\title{Create a float (continuous) distribution}
\usage{
float_distribution(low, high, log = FALSE, step = NULL)
}
\arguments{
\item{low}{Lower bound (inclusive).}

\item{high}{Upper bound (inclusive).}

\item{log}{If TRUE, sample in log space. Both bounds must be positive.}

\item{step}{If not NULL, round samples to this step size.}
}
\value{
An S3 object of class \code{roptuna_float_distribution}.
}
\description{
Create a float (continuous) distribution
}
\examples{
float_distribution(0, 1)
float_distribution(1e-5, 1e-1, log = TRUE)
}
