% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-grid.R
\name{grid_sampler}
\alias{grid_sampler}
\title{Create a grid sampler}
\usage{
grid_sampler(search_space)
}
\arguments{
\item{search_space}{Named list of choice vectors: \code{list(lr = c(0.01, 0.1), n = c(32L, 64L))}.}
}
\value{
A \code{GridSampler} R6 object.
}
\description{
Create a grid sampler
}
\examples{
sampler <- grid_sampler(list(lr = c(0.01, 0.1), depth = c(3L, 5L)))
study <- create_study(sampler = sampler)
}
