% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruner-hyperband.R
\name{HyperbandPruner}
\alias{HyperbandPruner}
\title{Hyperband pruner R6 class}
\description{
See \code{\link[=hyperband_pruner]{hyperband_pruner()}} for the recommended constructor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HyperbandPruner-new}{\code{HyperbandPruner$new()}}
\item \href{#method-HyperbandPruner-prune}{\code{HyperbandPruner$prune()}}
\item \href{#method-HyperbandPruner-clone}{\code{HyperbandPruner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HyperbandPruner-new"></a>}}
\if{latex}{\out{\hypertarget{method-HyperbandPruner-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperbandPruner$new(
  min_resource,
  reduction_factor,
  min_early_stopping_rate,
  n_brackets
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HyperbandPruner-prune"></a>}}
\if{latex}{\out{\hypertarget{method-HyperbandPruner-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperbandPruner$prune(study, trial_snap)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HyperbandPruner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HyperbandPruner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperbandPruner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
