% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage-memory.R
\name{InMemoryStorage}
\alias{InMemoryStorage}
\title{In-memory storage backend for roptuna studies}
\description{
In-memory storage backend for roptuna studies

In-memory storage backend for roptuna studies
}
\examples{
storage <- InMemoryStorage$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InMemoryStorage-new}{\code{InMemoryStorage$new()}}
\item \href{#method-InMemoryStorage-create_study}{\code{InMemoryStorage$create_study()}}
\item \href{#method-InMemoryStorage-get_study}{\code{InMemoryStorage$get_study()}}
\item \href{#method-InMemoryStorage-get_study_directions}{\code{InMemoryStorage$get_study_directions()}}
\item \href{#method-InMemoryStorage-set_study_user_attr}{\code{InMemoryStorage$set_study_user_attr()}}
\item \href{#method-InMemoryStorage-get_study_user_attrs}{\code{InMemoryStorage$get_study_user_attrs()}}
\item \href{#method-InMemoryStorage-create_trial}{\code{InMemoryStorage$create_trial()}}
\item \href{#method-InMemoryStorage-set_trial_state}{\code{InMemoryStorage$set_trial_state()}}
\item \href{#method-InMemoryStorage-set_trial_values}{\code{InMemoryStorage$set_trial_values()}}
\item \href{#method-InMemoryStorage-set_trial_param}{\code{InMemoryStorage$set_trial_param()}}
\item \href{#method-InMemoryStorage-set_trial_intermediate_value}{\code{InMemoryStorage$set_trial_intermediate_value()}}
\item \href{#method-InMemoryStorage-set_trial_user_attr}{\code{InMemoryStorage$set_trial_user_attr()}}
\item \href{#method-InMemoryStorage-get_trial}{\code{InMemoryStorage$get_trial()}}
\item \href{#method-InMemoryStorage-get_all_trials}{\code{InMemoryStorage$get_all_trials()}}
\item \href{#method-InMemoryStorage-find_study}{\code{InMemoryStorage$find_study()}}
\item \href{#method-InMemoryStorage-clone}{\code{InMemoryStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-create_study"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-create_study}{}}}
\subsection{Method \code{create_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$create_study(study_name, direction, directions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-get_study"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-get_study}{}}}
\subsection{Method \code{get_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$get_study(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-get_study_directions"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-get_study_directions}{}}}
\subsection{Method \code{get_study_directions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$get_study_directions(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-set_study_user_attr"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-set_study_user_attr}{}}}
\subsection{Method \code{set_study_user_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$set_study_user_attr(study_id, key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-get_study_user_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-get_study_user_attrs}{}}}
\subsection{Method \code{get_study_user_attrs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$get_study_user_attrs(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-create_trial"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-create_trial}{}}}
\subsection{Method \code{create_trial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$create_trial(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-set_trial_state"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-set_trial_state}{}}}
\subsection{Method \code{set_trial_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$set_trial_state(
  study_id,
  trial_id,
  state,
  value = NULL,
  datetime_complete = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-set_trial_values"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-set_trial_values}{}}}
\subsection{Method \code{set_trial_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$set_trial_values(study_id, trial_id, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-set_trial_param"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-set_trial_param}{}}}
\subsection{Method \code{set_trial_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$set_trial_param(
  study_id,
  trial_id,
  param_name,
  distribution,
  value
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-set_trial_intermediate_value"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-set_trial_intermediate_value}{}}}
\subsection{Method \code{set_trial_intermediate_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$set_trial_intermediate_value(study_id, trial_id, step, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-set_trial_user_attr"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-set_trial_user_attr}{}}}
\subsection{Method \code{set_trial_user_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$set_trial_user_attr(study_id, trial_id, key_name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-get_trial"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-get_trial}{}}}
\subsection{Method \code{get_trial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$get_trial(study_id, trial_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-get_all_trials"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-get_all_trials}{}}}
\subsection{Method \code{get_all_trials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$get_all_trials(study_id, states = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-find_study"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-find_study}{}}}
\subsection{Method \code{find_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$find_study(study_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InMemoryStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InMemoryStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InMemoryStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
