% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{int_distribution}
\alias{int_distribution}
\title{Create an integer distribution}
\usage{
int_distribution(low, high)
}
\arguments{
\item{low}{Lower bound (inclusive).}

\item{high}{Upper bound (inclusive).}
}
\value{
An S3 object of class \code{roptuna_int_distribution}.
}
\description{
Create an integer distribution
}
\examples{
int_distribution(1L, 10L)
}
