% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage-journal.R
\name{JournalStorage}
\alias{JournalStorage}
\title{Journal storage R6 class (file-based, append-only)}
\description{
See \code{\link[=journal_storage]{journal_storage()}} for the recommended constructor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JournalStorage-new}{\code{JournalStorage$new()}}
\item \href{#method-JournalStorage-create_study}{\code{JournalStorage$create_study()}}
\item \href{#method-JournalStorage-get_study}{\code{JournalStorage$get_study()}}
\item \href{#method-JournalStorage-get_study_directions}{\code{JournalStorage$get_study_directions()}}
\item \href{#method-JournalStorage-set_study_user_attr}{\code{JournalStorage$set_study_user_attr()}}
\item \href{#method-JournalStorage-get_study_user_attrs}{\code{JournalStorage$get_study_user_attrs()}}
\item \href{#method-JournalStorage-find_study}{\code{JournalStorage$find_study()}}
\item \href{#method-JournalStorage-create_trial}{\code{JournalStorage$create_trial()}}
\item \href{#method-JournalStorage-set_trial_state}{\code{JournalStorage$set_trial_state()}}
\item \href{#method-JournalStorage-set_trial_values}{\code{JournalStorage$set_trial_values()}}
\item \href{#method-JournalStorage-set_trial_param}{\code{JournalStorage$set_trial_param()}}
\item \href{#method-JournalStorage-set_trial_intermediate_value}{\code{JournalStorage$set_trial_intermediate_value()}}
\item \href{#method-JournalStorage-set_trial_user_attr}{\code{JournalStorage$set_trial_user_attr()}}
\item \href{#method-JournalStorage-get_trial}{\code{JournalStorage$get_trial()}}
\item \href{#method-JournalStorage-get_all_trials}{\code{JournalStorage$get_all_trials()}}
\item \href{#method-JournalStorage-clone}{\code{JournalStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$new(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-create_study"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-create_study}{}}}
\subsection{Method \code{create_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$create_study(study_name, direction, directions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-get_study"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-get_study}{}}}
\subsection{Method \code{get_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$get_study(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-get_study_directions"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-get_study_directions}{}}}
\subsection{Method \code{get_study_directions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$get_study_directions(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-set_study_user_attr"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-set_study_user_attr}{}}}
\subsection{Method \code{set_study_user_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$set_study_user_attr(study_id, key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-get_study_user_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-get_study_user_attrs}{}}}
\subsection{Method \code{get_study_user_attrs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$get_study_user_attrs(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-find_study"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-find_study}{}}}
\subsection{Method \code{find_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$find_study(study_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-create_trial"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-create_trial}{}}}
\subsection{Method \code{create_trial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$create_trial(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-set_trial_state"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-set_trial_state}{}}}
\subsection{Method \code{set_trial_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$set_trial_state(
  study_id,
  trial_id,
  state,
  value = NULL,
  datetime_complete = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-set_trial_values"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-set_trial_values}{}}}
\subsection{Method \code{set_trial_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$set_trial_values(study_id, trial_id, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-set_trial_param"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-set_trial_param}{}}}
\subsection{Method \code{set_trial_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$set_trial_param(
  study_id,
  trial_id,
  param_name,
  distribution,
  value
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-set_trial_intermediate_value"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-set_trial_intermediate_value}{}}}
\subsection{Method \code{set_trial_intermediate_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$set_trial_intermediate_value(study_id, trial_id, step, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-set_trial_user_attr"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-set_trial_user_attr}{}}}
\subsection{Method \code{set_trial_user_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$set_trial_user_attr(study_id, trial_id, key_name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-get_trial"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-get_trial}{}}}
\subsection{Method \code{get_trial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$get_trial(study_id, trial_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-get_all_trials"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-get_all_trials}{}}}
\subsection{Method \code{get_all_trials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$get_all_trials(study_id, states = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JournalStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JournalStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JournalStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
