% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-nsgaii.R
\name{NSGAIISampler}
\alias{NSGAIISampler}
\title{NSGA-II sampler R6 class}
\description{
See \code{\link[=nsgaii_sampler]{nsgaii_sampler()}} for the recommended constructor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NSGAIISampler-new}{\code{NSGAIISampler$new()}}
\item \href{#method-NSGAIISampler-infer_relative_search_space}{\code{NSGAIISampler$infer_relative_search_space()}}
\item \href{#method-NSGAIISampler-sample_relative}{\code{NSGAIISampler$sample_relative()}}
\item \href{#method-NSGAIISampler-sample_independent}{\code{NSGAIISampler$sample_independent()}}
\item \href{#method-NSGAIISampler-clone}{\code{NSGAIISampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSGAIISampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-NSGAIISampler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSGAIISampler$new(population_size, eta_c, eta_m, mutation_prob, seed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSGAIISampler-infer_relative_search_space"></a>}}
\if{latex}{\out{\hypertarget{method-NSGAIISampler-infer_relative_search_space}{}}}
\subsection{Method \code{infer_relative_search_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSGAIISampler$infer_relative_search_space(study, trial)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSGAIISampler-sample_relative"></a>}}
\if{latex}{\out{\hypertarget{method-NSGAIISampler-sample_relative}{}}}
\subsection{Method \code{sample_relative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSGAIISampler$sample_relative(study, trial, search_space, all_params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSGAIISampler-sample_independent"></a>}}
\if{latex}{\out{\hypertarget{method-NSGAIISampler-sample_independent}{}}}
\subsection{Method \code{sample_independent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSGAIISampler$sample_independent(study, trial, param_name, dist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSGAIISampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NSGAIISampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSGAIISampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
