% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-random.R
\name{random_sampler}
\alias{random_sampler}
\title{Create a random sampler}
\usage{
random_sampler(seed = NULL)
}
\arguments{
\item{seed}{Optional random seed.}
}
\value{
A \code{RandomSampler} R6 object.
}
\description{
Create a random sampler
}
\examples{
sampler <- random_sampler(seed = 42L)
study <- create_study(sampler = sampler)
}
