% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roptuna-package.R
\docType{package}
\name{roptuna-package}
\alias{roptuna}
\alias{roptuna-package}
\title{roptuna: Hyperparameter Optimization Using the Optuna Framework}
\description{
An R implementation of the Optuna hyperparameter optimization framework.
Provides define-by-run search spaces, TPE sampling, trial pruning, and
persistent study storage via SQLite.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kvenkita.github.io/roptuna}
  \item \url{https://github.com/kvenkita/roptuna}
  \item Report bugs at \url{https://github.com/kvenkita/roptuna/issues}
}

}
\author{
\strong{Maintainer}: Kailas Venkitasubramanian \email{kailasv@gmail.com}

}
\keyword{internal}
