% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage-sqlite.R
\name{SqliteStorage}
\alias{SqliteStorage}
\title{SQLite storage R6 class (schema matches Python Optuna)}
\description{
See \code{\link[=sqlite_storage]{sqlite_storage()}} for the recommended constructor.
}
\examples{
\dontrun{
storage <- SqliteStorage$new(tempfile(fileext = ".sqlite"))
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SqliteStorage-new}{\code{SqliteStorage$new()}}
\item \href{#method-SqliteStorage-create_study}{\code{SqliteStorage$create_study()}}
\item \href{#method-SqliteStorage-get_study}{\code{SqliteStorage$get_study()}}
\item \href{#method-SqliteStorage-create_trial}{\code{SqliteStorage$create_trial()}}
\item \href{#method-SqliteStorage-set_trial_state}{\code{SqliteStorage$set_trial_state()}}
\item \href{#method-SqliteStorage-set_trial_param}{\code{SqliteStorage$set_trial_param()}}
\item \href{#method-SqliteStorage-set_trial_intermediate_value}{\code{SqliteStorage$set_trial_intermediate_value()}}
\item \href{#method-SqliteStorage-set_trial_user_attr}{\code{SqliteStorage$set_trial_user_attr()}}
\item \href{#method-SqliteStorage-get_trial}{\code{SqliteStorage$get_trial()}}
\item \href{#method-SqliteStorage-get_all_trials}{\code{SqliteStorage$get_all_trials()}}
\item \href{#method-SqliteStorage-clone}{\code{SqliteStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$new(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-create_study"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-create_study}{}}}
\subsection{Method \code{create_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$create_study(study_name, direction)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-get_study"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-get_study}{}}}
\subsection{Method \code{get_study()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$get_study(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-create_trial"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-create_trial}{}}}
\subsection{Method \code{create_trial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$create_trial(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-set_trial_state"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-set_trial_state}{}}}
\subsection{Method \code{set_trial_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$set_trial_state(
  study_id,
  trial_id,
  state,
  value = NULL,
  datetime_complete = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-set_trial_param"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-set_trial_param}{}}}
\subsection{Method \code{set_trial_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$set_trial_param(
  study_id,
  trial_id,
  param_name,
  distribution,
  value
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-set_trial_intermediate_value"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-set_trial_intermediate_value}{}}}
\subsection{Method \code{set_trial_intermediate_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$set_trial_intermediate_value(study_id, trial_id, step, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-set_trial_user_attr"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-set_trial_user_attr}{}}}
\subsection{Method \code{set_trial_user_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$set_trial_user_attr(study_id, trial_id, key_name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-get_trial"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-get_trial}{}}}
\subsection{Method \code{get_trial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$get_trial(study_id, trial_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-get_all_trials"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-get_all_trials}{}}}
\subsection{Method \code{get_all_trials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$get_all_trials(study_id, states = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqliteStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SqliteStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqliteStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
