% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-mlr3.R
\name{TunerOptuna}
\alias{TunerOptuna}
\title{mlr3tuning adapter for roptuna}
\description{
mlr3tuning adapter for roptuna

mlr3tuning adapter for roptuna
}
\details{
A class that uses roptuna to search hyperparameters inside an mlr3tuning
instance. Use with \code{mlr3tuning::ti()} and \code{tuner$optimize(instance)}.
}
\examples{
\dontrun{
# Requires mlr3, mlr3tuning, paradox packages
library(mlr3); library(mlr3tuning); library(paradox)
tuner <- TunerOptuna$new(sampler = tpe_sampler(seed = 1L))
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TunerOptuna-new}{\code{TunerOptuna$new()}}
\item \href{#method-TunerOptuna-optimize}{\code{TunerOptuna$optimize()}}
\item \href{#method-TunerOptuna-clone}{\code{TunerOptuna$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerOptuna-new"></a>}}
\if{latex}{\out{\hypertarget{method-TunerOptuna-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerOptuna$new(sampler = tpe_sampler(), pruner = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampler}}{roptuna sampler (default: tpe_sampler()).}

\item{\code{pruner}}{roptuna pruner or NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerOptuna-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-TunerOptuna-optimize}{}}}
\subsection{Method \code{optimize()}}{
Run optimization on an mlr3tuning TuningInstance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerOptuna$optimize(inst)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerOptuna-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TunerOptuna-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerOptuna$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
