% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruner-wilcoxon.R
\name{WilcoxonPruner}
\alias{WilcoxonPruner}
\title{Wilcoxon pruner R6 class}
\description{
See \code{\link[=wilcoxon_pruner]{wilcoxon_pruner()}} for the recommended constructor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WilcoxonPruner-new}{\code{WilcoxonPruner$new()}}
\item \href{#method-WilcoxonPruner-prune}{\code{WilcoxonPruner$prune()}}
\item \href{#method-WilcoxonPruner-clone}{\code{WilcoxonPruner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WilcoxonPruner-new"></a>}}
\if{latex}{\out{\hypertarget{method-WilcoxonPruner-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WilcoxonPruner$new(p_threshold, n_startup_trials)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WilcoxonPruner-prune"></a>}}
\if{latex}{\out{\hypertarget{method-WilcoxonPruner-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WilcoxonPruner$prune(study, trial_snap)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WilcoxonPruner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WilcoxonPruner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WilcoxonPruner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
