% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{Study}
\alias{Study}
\title{Study R6 class for hyperparameter optimization}
\description{
Study R6 class for hyperparameter optimization

Study R6 class for hyperparameter optimization
}
\details{
Manages the optimization loop. Create via \code{\link[=create_study]{create_study()}}.
}
\examples{
study <- create_study(direction = "minimize")
study$optimize(function(trial) trial$suggest_float("x", -5, 5)^2, n_trials = 5)
study$best_value
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Study-new}{\code{Study$new()}}
\item \href{#method-Study-optimize}{\code{Study$optimize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-new"></a>}}
\if{latex}{\out{\hypertarget{method-Study-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$new(study_id, study_name, direction, sampler, pruner, storage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-Study-optimize}{}}}
\subsection{Method \code{optimize()}}{
Run the objective function for n_trials iterations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$optimize(func, n_trials = 100, timeout = NULL, catch = character())}\if{html}{\out{</div>}}
}

}
}
